fix linux pid kill

+++ src/index.ts
@@ -1,4 +1,4 @@
-import { spawn, exec, ChildProcessWithoutNullStreams } from "child_process";
+import { spawn, ChildProcessWithoutNullStreams } from "child_process";
 import path from "path";
 import { EventEmitter } from "events";
 import WebSocket, { WebSocketServer } from "ws";
@@ -37,11 +37,11 @@ export interface CycleTLSRequestOptions {
     [key: string]: any;
   };
   cookies?:
-     Array<object>
-    | {
-        [key: string]: string;
-      };
-  body?: string | URLSearchParams | FormData;  
+  Array<object>
+  | {
+    [key: string]: string;
+  };
+  body?: string | URLSearchParams | FormData;
   ja3?: string;
   userAgent?: string;
   proxy?: string;
@@ -70,8 +70,16 @@ const cleanExit = async (message?: string | Error, exit?: boolean) => {
   if (message) console.log(message);
   exit = exit ?? true;
 
-  child?.kill();
-  if (exit) process.exit();
+  if (child) {
+
+    if (process.platform == "win32") {
+      child?.kill();
+    } else {
+      process.kill(-child.pid);
+    }
+    if (exit) process.exit();
+
+  }
 };
 process.on("SIGINT", () => cleanExit());
 process.on("SIGTERM", () => cleanExit());
@@ -80,7 +88,7 @@ const handleSpawn = (debug: boolean, fileName: string, port: number, filePath?: 
   const execPath = filePath ? `"${filePath}"` : `"${path.join(__dirname, fileName)}"`;
   child = spawn(execPath, {
     env: { WS_PORT: port.toString() },
-    shell: false,
+    shell: true,
     windowsHide: true,
     detached: process.platform !== "win32"
   });
@@ -118,7 +126,7 @@ class Golang extends EventEmitter {
   queue: Array<string>;
   host: boolean;
   queueId: NodeJS.Timeout;
-  
+
   private timeout: number;
   private port: number;
   private debug: boolean;
@@ -136,28 +144,28 @@ class Golang extends EventEmitter {
     this.checkSpawnedInstance();
   }
 
-  checkSpawnedInstance(){
+  checkSpawnedInstance() {
     let server = http.createServer();
 
     server.listen(this.port)
-        .on('listening', () => {
-          server.close(() => {
-            this.spawnServer();
-            this.host = true;
-          })
+      .on('listening', () => {
+        server.close(() => {
+          this.spawnServer();
+          this.host = true;
         })
-        .on('error', () => {
-          this.createClient();
-          this.host = false;
-        });
+      })
+      .on('error', () => {
+        this.createClient();
+        this.host = false;
+      });
   }
 
-  spawnServer(){
+  spawnServer() {
     const PLATFORM_BINARIES: { [platform: string]: { [arch: string]: string } } = {
-      "win32":    { "x64": "index.exe" },
-      "linux":    { "arm": "index-arm", "arm64": "index-arm64", "x64": "index" },
-      "darwin":   { "x64": "index-mac", "arm": "index-mac-arm", "arm64": "index-mac-arm64" },
-      "freebsd":  { "x64": "index-freebsd" }
+      "win32": { "x64": "index.exe" },
+      "linux": { "arm": "index-arm", "arm64": "index-arm64", "x64": "index" },
+      "darwin": { "x64": "index-mac", "arm": "index-mac-arm", "arm64": "index-mac-arm64" },
+      "freebsd": { "x64": "index-freebsd" }
     };
 
     const executableFilename = PLATFORM_BINARIES[process.platform]?.[os.arch()];
@@ -170,7 +178,7 @@ class Golang extends EventEmitter {
     this.createClient();
   }
 
-  createClient(){
+  createClient() {
     // In-line function that represents a connection attempt
     const attemptConnection = () => {
       const server = new WebSocket(`ws://localhost:${this.port}`);
@@ -193,7 +201,7 @@ class Golang extends EventEmitter {
 
         setTimeout(() => {
           // If we've failed to initialize, stop the loop
-          if(this.failedInitialization){
+          if (this.failedInitialization) {
             return;
           }
 
@@ -234,8 +242,8 @@ class Golang extends EventEmitter {
 
         // First, we'll create a queue to store the failed request
         // Do a check to make sure server isn't null to prevent a race condition where multiple requests fail
-        if(err){
-          if(this.server != null){
+        if (err) {
+          if (this.server != null) {
             // Add failed request to queue
             this.server = null;
 
@@ -244,14 +252,14 @@ class Golang extends EventEmitter {
 
             // Start process of client re-creation
             this.checkSpawnedInstance();
-          }else{
+          } else {
             // Add to queue and hope server restarts properly
-            this.queue.push(JSON.stringify({requestId, options}));
+            this.queue.push(JSON.stringify({ requestId, options }));
           }
         }
       });
     } else {
-      if(this.queue == null){
+      if (this.queue == null) {
         this.queue = [];
       }
       this.queue.push(JSON.stringify({ requestId, options }))
@@ -259,7 +267,7 @@ class Golang extends EventEmitter {
       if (this.queueId == null) {
         this.queueId = setInterval(() => {
           // If we failed to initialize - clear the queue
-          if(this.failedInitialization){
+          if (this.failedInitialization) {
             clearInterval(this.queueId);
             this.queue = [];
             this.queueId = null;
@@ -284,7 +292,11 @@ class Golang extends EventEmitter {
     return new Promise((resolve, reject) => {
       this.server.close();
       if (this.host) {
-        child?.kill();
+        if (process.platform == "win32") {
+          child?.kill();
+        } else {
+          process.kill(-child.pid);
+        }
         resolve(null);
       } else {
         resolve(null);
@@ -294,9 +306,9 @@ class Golang extends EventEmitter {
 }
 export interface CycleTLSClient {
   (
-      url: string,
-      options: CycleTLSRequestOptions,
-      method?: "head" | "get" | "post" | "put" | "delete" | "trace" | "options" | "connect" | "patch"
+    url: string,
+    options: CycleTLSRequestOptions,
+    method?: "head" | "get" | "post" | "put" | "delete" | "trace" | "options" | "connect" | "patch"
   ): Promise<CycleTLSResponse>;
   head(url: string, options: CycleTLSRequestOptions): Promise<CycleTLSResponse>;
   get(url: string, options: CycleTLSRequestOptions): Promise<CycleTLSResponse>;
@@ -345,25 +357,25 @@ const initCycleTLS = async (
             const requestId = `${url}${Math.floor(Date.now() * Math.random())}`;
             //set default options
             options = options ?? {}
-            
+
             //set default ja3, user agent, body and proxy
             if (!options?.ja3)
               options.ja3 = "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-51-57-47-53-10,0-23-65281-10-11-35-16-5-51-43-13-45-28-21,29-23-24-25-256-257,0";
-              if (!options?.userAgent)
+            if (!options?.userAgent)
               options.userAgent = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/101.0.4951.54 Safari/537.36";
             if (!options?.body) options.body = "";
             if (!options?.proxy) options.proxy = "";
             if (!options?.insecureSkipVerify) options.insecureSkipVerify = false;
             if (!options?.forceHTTP1) options.forceHTTP1 = false;
 
-            
+
             //convert simple cookies
             const cookies = options?.cookies;
             if (
               typeof cookies === "object" &&
               !Array.isArray(cookies) &&
               cookies !== null
-             ) {
+            ) {
               const tempArr: {
                 [key: string]: any;
               } = [];
@@ -384,11 +396,11 @@ const initCycleTLS = async (
                 //parse json responses
                 response.Body = JSON.parse(response.Body);
                 //override console.log full repl to display full body
-                response.Body[util.inspect.custom] = function(){ return JSON.stringify( this, undefined, 2); }
-              } catch (e) {}
+                response.Body[util.inspect.custom] = function () { return JSON.stringify(this, undefined, 2); }
+              } catch (e) { }
 
               const { Status: status, Body: body, Headers: headers, FinalUrl: finalUrl } = response;
-              
+
               if (headers["Set-Cookie"])
                 headers["Set-Cookie"] = headers["Set-Cookie"].split("/,/");
               resolveRequest({
